error code is `u64` (#246)
* error code is u64

* fix tests

+++ crates/ide-tests/src/test_error_const_docs.rs
@@ -9,7 +9,7 @@ fn test_no_error_if_const_is_not_used_as_error_in_assert_or_abort() {
     // language=Move
     check_diagnostics(expect![[r#"
             module 0x1::m {
-                const ERR_ONE: u8 = 1;
+                const ERR_ONE: u64 = 1;
                 fun main() {
                     ERR_ONE + 1;
                 }
@@ -23,9 +23,9 @@ fn test_no_error_if_const_has_doc_comment() {
     check_diagnostics(expect![[r#"
             module 0x1::m {
                 /// docs
-                const ERR_ONE: u8 = 1;
+                const ERR_ONE: u64 = 1;
                 /// docs
-                const ERR_TWO: u8 = 1;
+                const ERR_TWO: u64 = 1;
                 fun main() {
                     assert!(true, ERR_ONE);
                     abort ERR_TWO;
@@ -39,7 +39,7 @@ fn test_no_error_path_has_colon_colon() {
     // language=Move
     check_diagnostics(expect![[r#"
         module 0x1::m {
-            const ERR_ONE: u8 = 1;
+            const ERR_ONE: u64 = 1;
             fun main() {
                 assert!(true, Self::ERR_ONE);
             }
@@ -52,7 +52,7 @@ fn test_error_no_doc_comment_assert() {
     // language=Move
     check_diagnostics(expect![[r#"
         module 0x1::m {
-            const ERR_ONE: u8 = 1;
+            const ERR_ONE: u64 = 1;
                 //^^^^^^^ warn: Missing documentation comment (provides a human-readable error message on-chain)
             fun main() {
                 assert!(true, ERR_ONE);
@@ -66,7 +66,7 @@ fn test_error_no_doc_comment_abort() {
     // language=Move
     check_diagnostics(expect![[r#"
         module 0x1::m {
-            const ERR_ONE: u8 = 1;
+            const ERR_ONE: u64 = 1;
                 //^^^^^^^ warn: Missing documentation comment (provides a human-readable error message on-chain)
             fun main() {
                 abort ERR_ONE;
@@ -81,7 +81,7 @@ fn test_error_no_doc_comment_simple_comment_assert() {
     check_diagnostics(expect![[r#"
         module 0x1::m {
             // not docs
-            const ERR_ONE: u8 = 1;
+            const ERR_ONE: u64 = 1;
                 //^^^^^^^ warn: Missing documentation comment (provides a human-readable error message on-chain)
             fun main() {
                 assert!(true, ERR_ONE);
